refactor queries into ProjectV2Content
